databricks-crud-app/requirements.txt
dash==3.2.0
psycopg-pool==3.2.6
psycopg-binary==3.2.10
psycopg==3.2.10
databricks-sdk==0.67.0
pandas==2.2.3
databricks-sql-connector==4.1.3
pyarrow==21.0.0